Got everything working for logging data.
Running the log_flight.py logs data onto SD card while flying an eight.
The function for flying in it can be changed to fly another route.
After it has finished the log can be decoded with decode_logs.py which
takes the filename as an argument and outputs the decded logs in the
decoded folder.

Next step is to test it for different flight patterns and controllers.
As well as writing some scripts to analyze the data.